Move skipped test
